+++ R/jsFunc-click.R
@@ -4,6 +4,8 @@
 #' on a Shiny \code{actionButton()}.
 #'
 #' @param id The id of the button
+#' @param asis If \code{TRUE}, use the ID as-is even when inside a module
+#' (instead of adding the namespace prefix to the ID).
 #' @seealso \code{\link[shinyjs]{useShinyjs}},
 #' \code{\link[shinyjs]{runExample}}
 #' @note \code{shinyjs} must be initialized with a call to \code{useShinyjs()}
@@ -31,8 +33,8 @@
 #'   )
 #' }
 #' @export
-click <- function(id) {
+click <- function(id, asis = FALSE) {
   fxn <- "click"
-  params <- list(id = id)
+  params <- list(id = id, asis = asis)
   jsFuncHelper(fxn, params)
 }